requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
boto3==1.33.13
botocore==1.33.13
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Django==3.2.25
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.1.0
django-redis==5.4.0
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.2.2
fastapi==0.103.2
h11==0.14.0
idna==3.7
importlib-metadata==6.7.0
jmespath==1.0.1
motor==3.4.0
outcome==1.3.0.post0
Pillow==9.5.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydub==0.25.1
PyJWT==2.8.0
pymongo==4.7.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
redis==5.0.8
requests==2.31.0
s3transfer==0.8.2
selenium==4.11.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SpeechRecognition==3.9.0
sqlparse==0.4.4
starlette==0.27.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
urllib3==1.26.20
uvicorn==0.22.0
wsproto==1.2.0
zipp==3.15.0
